--- packages/rx_bloc_cli/mason_templates/bricks/lib_auth_google/CHANGELOG.md
@@ -1,3 +1,0 @@
-# 0.1.0+1
-
-- TODO: Describe initial release.
