--- packages/rx_bloc_cli/mason_templates/bricks/lib_auth_google/brick.yaml
@@ -1,7 +1,0 @@
-name: lib_auth_google
-description: Adds new lib directory which contains all the layers needed to implement google login functionality
-
-version: 0.1.0+1
-environment:
-  mason: ">=0.1.0-dev.47 <0.1.0"
-vars:
