+++ packages/rx_bloc_cli/mason_templates/bricks/lib_social_logins/__brick__/lib/lib_social_logins/data_sources/google_auth_data_source.dart
@@ -1,10 +1,10 @@
-{{> licence.dart}}
+{{> licence.dart }}
 
 import 'package:dio/dio.dart';
 import 'package:retrofit/http.dart';
 
 import '../../../lib_auth/models/auth_token_model.dart';
-import '../../models/google_auth_request_model.dart';
+import '../models/google_auth_request_model.dart';
 
 part 'google_auth_data_source.g.dart';
