+++ packages/rx_bloc_cli/mason_templates/bricks/lib_social_logins/__brick__/lib/lib_social_logins/data_sources/google_credential_data_source.dart
@@ -1,4 +1,4 @@
-{{> licence.dart}}
+{{> licence.dart }}
 
 import 'package:google_sign_in/google_sign_in.dart';
