+++ packages/rx_bloc_cli/mason_templates/bricks/lib_social_logins/__brick__/lib/lib_social_logins/models/google_credentials_model.dart
@@ -1,8 +1,15 @@
 {{> licence.dart }}
 
+// ignore_for_file: public_member_api_docs, sort_constructors_first
 import 'package:google_sign_in/google_sign_in.dart';
 
 class GoogleCredentialsModel {
+  final String? displayName;
+  final String email;
+  final String id;
+  final String? photoUrl;
+  final String? serverAuthCode;
+
   GoogleCredentialsModel({
     this.displayName,
     required this.email,
@@ -10,12 +17,6 @@ class GoogleCredentialsModel {
     this.photoUrl,
     this.serverAuthCode,
   });
-  
-  final String? displayName;
-  final String email;
-  final String id;
-  final String? photoUrl;
-  final String? serverAuthCode;
 
   bool equals(GoogleCredentialsModel credentials) =>
       displayName == credentials.displayName &&
