+++ packages/rx_bloc_cli/mason_templates/bricks/lib_social_logins/__brick__/lib/lib_social_logins/repositories/google_auth_repository.dart
@@ -4,8 +4,8 @@ import '../../assets.dart';
 import '../../base/common_mappers/error_mappers/error_mapper.dart';
 import '../../base/models/errors/error_model.dart';
 import '../../lib_auth/models/auth_token_model.dart';
-import '../data_sources/remote/google_auth_data_source.dart';
-import '../data_sources/remote/google_credential_data_source.dart';
+import '../data_sources/google_auth_data_source.dart';
+import '../data_sources/google_credential_data_source.dart';
 import '../models/google_auth_request_model.dart';
 import '../models/google_credentials_model.dart';
