+++ packages/rx_bloc_cli/mason_templates/bricks/lib_social_logins/__brick__/lib/lib_social_logins/services/apple_social_login_service.dart
@@ -1,10 +1,9 @@
 {{> licence.dart }}
 
 import '../../lib_auth/models/auth_token_model.dart';
-import 'social_login_service.dart';
 import '../../lib_auth/services/user_account_service.dart';
-import '../models/apple_credential_model.dart';
 import '../repositories/apple_auth_repository.dart';
+import 'social_login_service.dart';
 
 class AppleSocialLoginService extends SocialLoginService {
   AppleSocialLoginService(
