+++ packages/rx_bloc_cli/mason_templates/bricks/lib_social_logins/__brick__/lib/lib_social_logins/services/google_login_service.dart
@@ -1,10 +1,10 @@
 {{> licence.dart }}
 
 import '../../lib_auth/models/auth_token_model.dart';
-import '../../lib_auth/services/social_login_service.dart';
 import '../../lib_auth/services/user_account_service.dart';
 import '../models/google_credentials_model.dart';
 import '../repositories/google_auth_repository.dart';
+import 'social_login_service.dart';
 
 class GoogleLoginService extends SocialLoginService {
   GoogleLoginService(
@@ -16,7 +16,6 @@ class GoogleLoginService extends SocialLoginService {
   Future<AuthTokenModel> authenticate() async {
     final GoogleCredentialsModel googleUser =
         await _googleAuthRepository.getUsersGoogleCredential();
-
     return _googleAuthRepository.googleAuth(googleAuthRequestModel: googleUser);
   }
 }
