+++ packages/rx_bloc_cli/mason_templates/bricks/rx_bloc_base/__brick__/lib/feature_login/views/login_page.dart
@@ -4,7 +4,8 @@ import 'package:flutter/material.dart';
 
 import '../../app_extensions.dart';
 import '../../base/common_ui_components/custom_app_bar.dart';{{#enable_social_logins}}
-import '../../lib_social_logins/ui_components/apple_login_widget.dart';{{/enable_social_logins}}
+import '../../lib_social_logins/ui_components/apple_login_widget.dart';
+import '../../lib_social_logins/ui_components/google_login_widget.dart';{{/enable_social_logins}}
 import '../ui_components/login_form.dart';
 
 class LoginPage extends StatelessWidget {
@@ -42,7 +43,15 @@ class LoginPage extends StatelessWidget {
                     vertical: context.designSystem.spacing.l,
                 ),
                 child: const AppleLoginWidget(),
-              ),{{/enable_social_logins}}
+              ),
+              Padding(
+                padding: EdgeInsets.symmetric(
+                    horizontal: context.designSystem.spacing.xxl,
+                    vertical: context.designSystem.spacing.l,
+                ),
+                child: const GoogleLoginWidget(),
+              ),
+              {{/enable_social_logins}}
             ],
           ),
         ),
