+++ packages/rx_bloc_cli/mason_templates/bricks/rx_bloc_base/__brick__/lib/l10n/arb/error/en.arb
@@ -12,7 +12,7 @@
   "passwordLength": "Password should be at least 6 characters long",
   "wrongEmailOrPassword": "Wrong email or password",
   "invalidMessage": "The entered message should be at least 2 characters long.",
-  "tooLong": "The entered text is too long",{{#enable_google_auth}}
-  "googleAuthError":"Google authorization failed",{{/enable_google_auth}}
+  "tooLong": "The entered text is too long",{{#enable_social_logins}}
+  "googleAuthError":"Google authorization failed",{{/enable_social_logins}}
   "tooShort": "The entered text is too short"
 }
