+++ packages/rx_bloc_cli/mason_templates/bricks/rx_bloc_base/__brick__/lib/l10n/arb/es.arb
@@ -10,7 +10,7 @@
     "pageWithResult" : "Enter message",
     "tryAgain" : "Try again",
     "confirm" : "Confirm",
-    "submit" : "Submit",{{#enable_google_auth}}
-    "googleLogin":"Google Log in",{{/enable_google_auth}}
+    "submit" : "Submit",{{#enable_social_logins}}
+    "googleLogin":"Google Log in",{{/enable_social_logins}}
     "navProfile" : "Profile"
 }
