+++ packages/rx_bloc_cli/mason_templates/bricks/rx_bloc_base/__brick__/pubspec.yaml
@@ -20,10 +20,11 @@ dependencies:
   flutter_localizations:
     sdk: flutter
   flutter_rx_bloc: ^5.0.2
-  flutter_secure_storage: ^7.0.1
+  flutter_secure_storage: ^7.0.1{{#enable_social_logins}}
+  flutter_signin_button: ^2.0.0{{/enable_social_logins}}
   go_router: ^6.2.0
-  golden_toolkit: ^0.14.0{{#enable_google_auth}}
-  google_sign_in: ^6.0.2{{/enable_google_auth}}
+  golden_toolkit: ^0.14.0{{#enable_social_logins}}
+  google_sign_in: ^6.0.2{{/enable_social_logins}}
   http: ^0.13.5 # This package is used for the local server (if not used, can be removed)
   jaguar_jwt: ^3.0.0 # This package is used for the local server (if not used, can be removed)
   json_annotation: ^4.7.0
